refactor F1-10: replace CapitalManager with three inbound ports
Split single CapitalManager interface into ReserveCapitalPort (sync),
ConfirmExecutionPort and ReleaseMarginPort (async). Each port has a single
method and its own use case: ReserveCapitalUseCase (full logic), and stubs
for ConfirmExecution/ReleaseMargin (scheduled for F1-11).

Update docs/IMPLEMENTATION_GUIDE.md to reflect 3-port architecture across
all references (Section 5.2, 5.3, diagrams, state table, tradeoffs).
